Order implemention done

+++ backend/src/main/java/com/pos/staff/dao/OrderDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pos.staff.entity.Order;
 
-public interface OrderDao extends JpaRepository<Order, Long > {
+public interface OrderDao extends JpaRepository<Order, Integer > {
 	
 
 }
